Finished refactoring documents
